requirements.txt
Django>=5.2
djangorestframework
pytest
pytest-asyncio
